// types.ts
export interface Psychographics {
  personalityTraits: string[];
  values: string[];
  painPoints: string[];
  motivations: string[];
}

export interface ViewerArchetype {
  name: string;
  age: number;
  profession: string;
  interests: string[];
  motivation: string;
}

export interface CompetitorAudience {
  channelName: string;
  similarities: string[];
  differences: string[];
}

export interface CommunityHotspot {
    platform: string;
    community: string;
    reason: string;
}

export interface AudienceProfileData {
  viewerSummary: {
    age: string;
    gender: string;
    country: string;
  };
  contentResonanceScore: number;
  psychographics: Psychographics;
  sentimentAndEmotion: {
    emotionalTriggers: string[];
    toneAlignment: string;
  };
  communityHotspots: CommunityHotspot[];
  behavioralPredictions: {
    activeHours: string;
    ctaResponsiveness: string;
    preferredFormats: string[];
  };
  viewerArchetypes: ViewerArchetype[];
  competitorAudienceOverlap: CompetitorAudience[];
  engagementTips: string[];
}